+++ app/src/main/java/de/baumann/hhsmoodle/fragmentsMain/FragmentInfo.java
@@ -164,6 +164,7 @@ public class FragmentInfo extends Fragment {
                                     shortcut.putExtra(Intent.EXTRA_SHORTCUT_ICON_RESOURCE, Intent.ShortcutIconResource.fromContext(getActivity().getApplicationContext(), R.mipmap.ic_launcher));
                                     shortcut.setAction("com.android.launcher.action.INSTALL_SHORTCUT");
                                     getActivity().sendBroadcast(shortcut);
+                                    Snackbar.make(listView, R.string.toast_shortcut, Snackbar.LENGTH_LONG).show();
                                 }